Refactor DeleteProfile component integration and enhance delete confirmation flow
