+++ client/src/components/DeleteProfile/DeleteProfile.jsx
@@ -1,14 +1,8 @@
-import { useState } from "react";
 import { UseUser } from "../../context/UserContext";
 import "./DeleteProfile.css";
 
-export default function DeleteProfile() {
-  const [showDeletePopup, setShowDeletePopup] = useState(false);
+export default function DeleteProfile({ setShowDeletePopup }) {
   const { deleteUser, dispatch, defaultUser } = UseUser();
-
-  const handleDeleteClick = () => {
-    setShowDeletePopup(true);
-  };
 
   const handleConfirmDelete = async () => {
     try {
@@ -29,41 +23,22 @@ export default function DeleteProfile() {
   };
 
   return (
-    <>
-      <div className="profile-delete-row">
-        <div>
-          <h3 className="profile-delete-title">Delete profile</h3>
-          <p className="profile-delete-desc">
-            Permanently delete your account and data.
-          </p>
+    <div className="profile-popup-overlay">
+      <div className="profile-popup-card">
+        <h2>Are you sure?</h2>
+        <p>This will permanently delete your account.</p>
+        <div className="profile-popup-buttons">
+          <button
+            className="profile-btn-secondary"
+            onClick={handleCancelDelete}
+          >
+            Cancel
+          </button>
+          <button className="profile-btn-primary" onClick={handleConfirmDelete}>
+            OK
+          </button>
         </div>
-        <button onClick={handleDeleteClick} className="profile-delete-btn">
-          Delete profile
-        </button>
       </div>
-      {/* DELETE CONFIRM POPUP */}
-      {showDeletePopup && (
-        <div className="profile-popup-overlay">
-          <div className="profile-popup-card">
-            <h2>Are you sure?</h2>
-            <p>This will permanently delete your account.</p>
-            <div className="profile-popup-buttons">
-              <button
-                className="profile-btn-secondary"
-                onClick={handleCancelDelete}
-              >
-                Cancel
-              </button>
-              <button
-                className="profile-btn-primary"
-                onClick={handleConfirmDelete}
-              >
-                OK
-              </button>
-            </div>
-          </div>
-        </div>
-      )}
-    </>
+    </div>
   );
 }
